Score Updation
I am displaying total score instead of the current score at the top. So, now I use a new variable total_score

Further, I am displaying current score at the end of each gait cycle. This was a little tricky. I display the score for a count of "display_score_max" after the cycle is complete. I do this because we can know the final score only after the cycle is completed.