Changes
1. Game
The file had some redundancy (data access from excewl sheet for dorsiflexion which we are not doing). Just removed the part of the code to make it look cleaner.

2. main_menu
Previously, only the game was modified. Now, the code with the main menu is modified to work with netcat channeled inputs.